wip – confList
