+++ dashboard-frontend/src/helpers/getField.ts
@@ -1,4 +1,5 @@
-import { FieldState, EventInterface } from '../types/form';
+import { FieldState } from '../types/conf';
+import { EventInterface } from '../types/form';
 
 export const getField = (fields: FieldState[], event: EventInterface) => {
   const index = fields.findIndex(
